src/requirements.txt
cryptography
python-gnupg
pyyaml
sopsy

#### Sources ####
# Vaultwarden
pykeepass

#### Destinations ####
# GCP
google-cloud-storage

# AWS